requirements.txt
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.22.0
pandas>=1.4.0
tqdm>=4.64.0
faiss-cpu>=1.7.2
scikit-learn>=1.0.2
langchain>=0.2.1
langchain-core>=0.2.1
langchain-community>=0.2.1
langchain-openai>=0.1.7
langchain-text-splitters>=0.2.2
chromadb>=0.5.0
openai>=1.30.0
